Refactor fixture matrix surface evidence parsing
