+++ lib/fixture-matrix/collectors/run-intake.mjs
@@ -41,6 +41,7 @@ import {
   normalizeLiveWpParityCollectorOptions,
 } from './live-wp-parity.mjs';
 import { normalizeFixtureMatrixResult, normalizeFixtureResult } from '../result.mjs';
+import { surfaceFields } from '../shared/surface-evidence.mjs';
 
 export function collectFixtureMatrixRunResults(input = {}) {
   const matrix = input.matrix || createFixtureMatrix(input);
@@ -670,17 +671,6 @@ function visitRuntimePayloads(value, inheritedFixtureId, payloads, seen, inherit
   for (const child of Object.values(value)) {
     visitRuntimePayloads(child, fixtureId, payloads, seen, surface);
   }
-}
-
-function surfaceFields(payload) {
-  const metadata = objectValue(payload.metadata || payload.recipeStepMetadata || payload.recipe_step_metadata);
-  return compactObject({
-    surface_id: payload.surface_id || payload.surfaceId || metadata.surface_id || metadata.surfaceId,
-    surface_label: payload.surface_label || payload.surfaceLabel || metadata.surface_label || metadata.surfaceLabel,
-    source_entry: payload.source_entry || payload.sourceEntry || metadata.source_entry || metadata.sourceEntry,
-    target: payload.target || metadata.target,
-    url: payload.url || payload.candidate_url || payload.candidateUrl || metadata.url || metadata.candidate_url || metadata.candidateUrl,
-  });
 }
 
 function surfaceSortKey(surface) {
