+++ lib/fixture-matrix/result.mjs
@@ -55,6 +55,12 @@ import {
   renderVisualParityEvidenceReportMarkdown,
 } from './visual-evidence-report.mjs';
 import {
+  hasEditorSurfaceEvidence,
+  hasVisualSurfaceEvidence,
+  surfaceArtifactPaths,
+  surfaceFields,
+} from './shared/surface-evidence.mjs';
+import {
   collectBlockComposition,
   computeFixtureEditorQuality,
   attachFixtureEditorQuality,
@@ -449,12 +455,13 @@ export function normalizeFixtureResult(input) {
 
 function normalizeSurfaceResult(input) {
   const row = objectValue(input);
+  const fields = surfaceFields(row, { includeGenericAliases: true });
   return compactObject({
-    surface_id: row.surface_id || row.surfaceId || row.id || '',
-    surface_label: row.surface_label || row.surfaceLabel || row.label || '',
-    source_entry: row.source_entry || row.sourceEntry || '',
-    target: row.target || '',
-    url: row.url || row.candidate_url || row.candidateUrl || '',
+    surface_id: fields.surface_id || '',
+    surface_label: fields.surface_label || '',
+    source_entry: fields.source_entry || '',
+    target: fields.target || '',
+    url: fields.url || '',
     artifact_refs: normalizeArray(row.artifact_refs || row.artifactRefs),
     editor_validation: row.editor_validation || row.editorValidation || null,
     editor_canvas: boundBlob(row.editor_canvas || row.editorCanvas || null),
@@ -611,24 +618,6 @@ function aggregateSurfaceEvidence(fixtures) {
       editor_artifact_paths: surfaceArtifactPaths(surface).filter((artifactPath) => /editor/i.test(artifactPath)),
     })),
   };
-}
-
-function hasVisualSurfaceEvidence(surface) {
-  const artifacts = objectValue(surface.visual_parity_artifacts || surface.visualParityArtifacts);
-  return Object.keys(objectValue(artifacts.artifacts)).length > 0
-    || Object.keys(objectValue(artifacts.metrics || artifacts.comparison)).length > 0
-    || surfaceArtifactPaths(surface).some((artifactPath) => /visual|screenshot|diff|\.png$/i.test(artifactPath));
-}
-
-function hasEditorSurfaceEvidence(surface) {
-  return Boolean(surface.editor_validation || surface.editorValidation || surface.editor_canvas || surface.editorCanvas || surface.editor_open || surface.editorOpen)
-    || surfaceArtifactPaths(surface).some((artifactPath) => /editor/i.test(artifactPath));
-}
-
-function surfaceArtifactPaths(surface) {
-  return normalizeArray(surface.artifact_refs || surface.artifactRefs)
-    .map((ref) => objectValue(ref).path || objectValue(ref).file || objectValue(ref).href || '')
-    .filter(Boolean);
 }
 
 function shouldStageFixtureSources(input = {}) {
